refactor(index): message event
