+++ index.js
@@ -1,7 +1,7 @@
 import { Client, Collection, Intents } from "discord.js";
 import { config } from "./utils/config.js";
-import { i18n } from "./utils/i18n.js";
 import { importCommands } from "./utils/importCommands.js";
+import { messageCreate } from "./utils/messageCreate.js";
 
 const { TOKEN, PREFIX } = config;
 
@@ -21,17 +21,13 @@ client.commands = new Collection();
 client.prefix = PREFIX;
 client.queue = new Map();
 
-const cooldowns = new Collection();
-const escapeRegex = (str) => str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-
 /**
- * Client Events
+ * Client events
  */
 client.on("ready", () => {
   console.log(`${client.user.username} ready!`);
   client.user.setActivity(`${PREFIX}help and ${PREFIX}play`, { type: "LISTENING" });
 });
-
 client.on("warn", (info) => console.log(info));
 client.on("error", console.error);
 
@@ -40,48 +36,7 @@ client.on("error", console.error);
  */
 importCommands(client);
 
-client.on("messageCreate", async (message) => {
-  if (message.author.bot || !message.guild) return;
-
-  const prefixRegex = new RegExp(`^(<@!?${client.user.id}>|${escapeRegex(PREFIX)})\\s*`);
-  if (!prefixRegex.test(message.content)) return;
-
-  const [, matchedPrefix] = message.content.match(prefixRegex);
-
-  const args = message.content.slice(matchedPrefix.length).trim().split(/ +/);
-  const commandName = args.shift().toLowerCase();
-
-  const command =
-    client.commands.get(commandName) || client.commands.find((cmd) => cmd.aliases?.includes(commandName));
-
-  if (!command) return;
-
-  if (!cooldowns.has(command.name)) {
-    cooldowns.set(command.name, new Collection());
-  }
-
-  const now = Date.now();
-  const timestamps = cooldowns.get(command.name);
-  const cooldownAmount = (command.cooldown || 1) * 1000;
-
-  if (timestamps.has(message.author.id)) {
-    const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
-
-    if (now < expirationTime) {
-      const timeLeft = (expirationTime - now) / 1000;
-      return message.reply(
-        i18n.__mf("common.cooldownMessage", { time: timeLeft.toFixed(1), name: command.name })
-      );
-    }
-  }
-
-  timestamps.set(message.author.id, now);
-  setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
-
-  try {
-    command.execute(message, args);
-  } catch (error) {
-    console.error(error);
-    message.reply(i18n.__("common.errorCommand")).catch(console.error);
-  }
-});
+/**
+ * Message event
+ */
+messageCreate(client);
